Model and Generator
also the initial generated railsy stuff.

The model is pretty straight forward stuff.  The email checking could be more in
depth but, this is what's in the rails docs and I didn't want to spend too much
time.  The generator exists to avoid using the AR lifecycle callbacks.  This
makes what's going on more explicit see :
http://www.reinteractive.net/posts/130-activerecord-callbacks-considered-harmful
and
http://www.samuelmullen.com/2013/05/the-problem-with-rails-callbacks/
for more in depth reasoning why AR lifecycle callbacks are a PITA